feat(tests) #add tests

+++ WebApiCarProject/Controllers/CarManagementController.cs
@@ -23,13 +23,13 @@ namespace WebApiCarProject.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetCar(int id)
+        public async Task<IActionResult> GetCar(long id)
         {
             var car = await _carRepository.GetCarAsync(id);
-            
+
             if (car == null)
                 return NotFound();
-            
+
             return Ok(car);
         }
 
@@ -38,35 +38,35 @@ namespace WebApiCarProject.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            
+
             await _carRepository.InsertCarAsync(car);
             await _carRepository.SaveAsync();
-            
+
             return CreatedAtAction("GetCar", new { id = car.Id }, car);
         }
-        
+
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutCar(int id, [FromBody] Car car)
+        public async Task<IActionResult> PutCar(long id, [FromBody] Car car)
         {
             if (id != car.Id)
                 return BadRequest();
-            
+
             //todo update logic.
 
-            await _carRepository.SaveAsync(); 
+            await _carRepository.SaveAsync();
             return NoContent();
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteCar(int id)
+        public async Task<IActionResult> DeleteCar(long id)
         {
             var car = await _carRepository.GetCarAsync(id);
             if (car == null)
                 return NotFound();
-            
+
             await _carRepository.DeleteCarAsync(id);
             await _carRepository.SaveAsync();
-            
+
             return NoContent();
         }
     }
