+++ WebApiCarProject/Infrastructure/Repositories/ICarRepository.cs
@@ -4,11 +4,11 @@ namespace WebApiCarProject.Infrastructure.Repositories
 {
     public interface ICarRepository
     {
-        Task<Car> GetCarAsync(int id);
+        Task<Car> GetCarAsync(long id);
         Task<IEnumerable<Car>> GetAllCarsAsync();
         void InsertCar(Car car);
         Task InsertCarAsync(Car car);
-        Task DeleteCarAsync(int carId);
+        Task DeleteCarAsync(long carId);
         void Save();
         Task SaveAsync();
     }
